requirements.txt
confluent-kafka~=2.10.0
pydantic-settings~=2.9.1
psycopg2-binary~=2.9.10